Check if AX_CHECK_COMPILE_FLAG is present.
AX_CHECK_COMPILE_FLAG is usually part of separate autoconf package,
called autoconf-archive, and on some distributions is not installed by
default.

When this macro is not present but user generates configure script,
it will complain with not so obvious message.

This commit fixes this by emitting clear message what is missing, so
user knows what to do next.