Special handling when the last entry is deleted on deactivation
When a device is deactivated, the entry is deleted from the list.
But the stepwise deactivation still goes on. But while the device
is still in the process to reach min_state, if another trip
activates the same device again, it should take precedence.
In this way it is not limited by the current state.