Add an explicit warning if a sensor is created without control_path.
Without this, the output will be an ominous warning:

    [WARN]sensor id 13 : No temp sysfs for reading raw temp

The extra space after `13` is meant to separate the index from the path,
but no path is supplied at all.

Changing this to warn at creation makes it easier to figure out what's
going on:

    [WARN]Missing control_path for sensor id 13 (type: amdgpu-temperature)
    [WARN]sensor id 13 : No temp sysfs for reading raw temp

And so it scares less, as I don't have an AMD GPU at all.